pyproject.toml
[tool.poetry]
package-mode = false
name = "spi-platform"
version = "0.1.0"
description = ""
authors = ["Вячеслав Чертан <slavakemdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.3"
django-ninja = "^1.6.2"
django-stubs = "^6.0.1"
orjson = "^3.11.7"
requests = "^2.32.5"
beautifulsoup4 = "^4.14.3"
playwright = "^1.58.0"
google-genai = "^1.68.0"
python-dotenv = "^1.2.2"
pillow = "^12.1.1"


[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
